pyproject.toml
[project]
name = "intershoprag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-ollama>=1.0.1",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "streamlit>=1.56.0",
    "trafilatura>=2.0.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
